chore: Move color transformers to new file
